More refactoring of cognito components
