--- frontend/src/cognito/components/Center/index.ts
@@ -1,1 +1,0 @@
-export { Center } from './Center'
