+++ frontend/src/cognito/components/LoadingMessage/LoadingMessage.tsx
@@ -1,7 +1,7 @@
-import { Box, Typography } from '@mui/material'
 import React from 'react'
+import { Box, Typography } from '@mui/material'
+import { fontSizes } from '../../../styling'
 import { Icon } from '../../../components/Icon'
-import { fontSize, colors } from '../../styles/variables'
 
 export type LoadingMessageProps = {
   message?: string
@@ -10,7 +10,7 @@ export type LoadingMessageProps = {
 export function LoadingMessage({ message = 'Loading...', ...props }: LoadingMessageProps): JSX.Element {
   return (
     <Box {...props}>
-      <Box component="div" fontSize={fontSize.xxl} p={1} mr={1}>
+      <Box component="div" fontSize={fontSizes.xxl} p={1} mr={1}>
         <Icon name="spinner-third" spin />
       </Box>
       <Typography variant="body2">{message}</Typography>
