+++ frontend/src/cognito/components/LoadingScreen/LoadingScreen.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
-import { Center } from '../Center'
 import { LoadingMessage, LoadingMessageProps } from '../LoadingMessage/LoadingMessage'
-import { fontSize, colors } from '../../styles/variables'
+import { Body } from '../../../components/Body'
 
 export type LoadingScreenProps = LoadingMessageProps & {
   message?: string
@@ -9,8 +8,8 @@ export type LoadingScreenProps = LoadingMessageProps & {
 
 export function LoadingScreen({ message = 'Loading...' }: LoadingScreenProps): JSX.Element {
   return (
-    <Center textAlign="center" color={colors.grayLight}>
+    <Body center>
       <LoadingMessage message={message} />
-    </Center>
+    </Body>
   )
 }
