+++ frontend/src/cognito/components/PasswordStrengthInput/PasswordStrengthInput.css
@@ -1,4 +1,4 @@
-@import '../../styles/variables.css';
+@import '../../../styling.css';
 .special-chars {
   white-space: normal !important;
 }
