--- frontend/src/pages/ConnectionsPage/index.ts
@@ -1,1 +1,0 @@
-export { ConnectionsPage } from './ConnectionsPage'
